--- src/api/java/de/fearnixx/jeak/service/controller/IResponseEntity.java
@@ -1,13 +1,0 @@
-package de.fearnixx.jeak.service.controller;
-
-import java.util.Map;
-
-public interface IResponseEntity<T> {
-    void addHeader(String header, String value);
-    void removeHeader(String header);
-    Map<String, String> getHeaders();
-    void setEntity(T entity);
-    T getEntity();
-    void setStatus(int status);
-    int getStatus();
-}
