--- src/api/java/de/fearnixx/jeak/service/controller/RequestMethod.java
@@ -1,10 +1,0 @@
-package de.fearnixx.jeak.service.controller;
-
-public enum RequestMethod {
-    POST,
-    PATCH,
-    PUT,
-    GET,
-    HEAD,
-    DELETE
-}
