--- src/api/java/de/fearnixx/jeak/service/controller/request/IRequestContext.java
@@ -1,36 +1,0 @@
-package de.fearnixx.jeak.service.controller.request;
-
-import org.eclipse.jetty.http.HttpStatus;
-
-import java.util.Optional;
-
-public interface IRequestContext {
-
-    /**
-     * Optionally retrieves a request attribute from the context.
-     * For officially supported attributes, see {@link Attributes}.
-     */
-    <T> Optional<T> optAttribute(String name, Class<T> hint);
-
-    final class Attributes {
-        /**
-         * {@link de.fearnixx.jeak.service.controller.request.token.IAuthenticationToken}
-         *
-         * @apiNote Optionally filled, if authentication is successful AND the "Token" authentication scheme is used.
-         *
-         * @implNote Required (non-optional) parameter injections cause {@link HttpStatus#UNAUTHORIZED_401} on unsuccessful authentication.
-         */
-        public static final String AUTHENTICATION_TOKEN = "auth:token:authenticationToken";
-
-        /**
-         * {@link de.fearnixx.jeak.teamspeak.data.IUser}
-         *
-         * @apiNote Optionally filled, if authentication is successful AND the subject is an user.
-         * @implNote Required (non-optional) parameter injections cause {@link HttpStatus#UNAUTHORIZED_401} on unsuccessful authentication or {@link HttpStatus#FORBIDDEN_403} for principals that aren't users.
-         */
-        public static final String AUTHENTICATION_USER = "auth:subject:authenticatedUser";
-
-        private Attributes() {
-        }
-    }
-}
