--- src/api/java/de/fearnixx/jeak/service/controller/request/token/IAuthenticationToken.java
@@ -1,16 +1,0 @@
-package de.fearnixx.jeak.service.controller.request.token;
-
-import de.fearnixx.jeak.service.permission.base.ISubject;
-import de.fearnixx.jeak.teamspeak.data.IUser;
-
-import java.time.ZonedDateTime;
-import java.util.Optional;
-
-public interface IAuthenticationToken {
-
-    String getTokenString();
-
-    IUser getTokenOwner();
-
-    Optional<ZonedDateTime> getExpiry();
-}
