+++ src/api/java/de/fearnixx/jeak/service/http/RequestMethod.java
@@ -1,0 +1,10 @@
+package de.fearnixx.jeak.service.http;
+
+public enum RequestMethod {
+    POST,
+    PATCH,
+    PUT,
+    GET,
+    HEAD,
+    DELETE
+}
