+++ src/api/java/de/fearnixx/jeak/service/http/exceptions/RegisterControllerException.java
@@ -1,0 +1,15 @@
+package de.fearnixx.jeak.service.http.exceptions;
+
+public class RegisterControllerException extends RuntimeException{
+    public RegisterControllerException(String message) {
+        super(message);
+    }
+
+    public RegisterControllerException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public RegisterControllerException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}
