+++ src/main/java/de/fearnixx/jeak/service/http/controller/IncapableDummyAdapter.java
@@ -1,0 +1,29 @@
+package de.fearnixx.jeak.service.http.controller;
+
+import de.fearnixx.jeak.Main;
+import de.fearnixx.jeak.service.http.connection.HttpServer;
+import de.fearnixx.jeak.service.http.connection.RestConfiguration;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class IncapableDummyAdapter extends HttpServer {
+
+    public static final boolean EXPERIMENTAL_REST_ENABLED = Main.getProperty("jeak.experimental.enable_rest", false);
+
+    private static final Logger logger = LoggerFactory.getLogger(IncapableDummyAdapter.class);
+
+    public IncapableDummyAdapter(RestConfiguration restConfiguration) {
+        super(restConfiguration);
+        logger.warn("REST functionality is experimental and therefore disabled by default. Enable with \"jeak.experimental.enable_rest\"");
+    }
+
+    @Override
+    public void start() {
+        logger.warn("Start doesnt work since REST functionality is experimental and therefore disabled by default. Enable with \"jeak.experimental.enable_rest\"");
+    }
+
+    @Override
+    public void registerController(ControllerContainer controllerContainer) {
+        // This adapter is just a dummy until the REST service has been released.
+    }
+}
